remove profiler, and just put it in flamegraph
