+++ lib/singed/kernel_ext.rb
@@ -1,6 +1,6 @@
 module Kernel
   def flamegraph(label = nil, open: true, ignore_gc: false, interval: 1000, io: $stdout, &)
-    fg = Singed::Flamegraph.new(label: label, ignore_gc: ignore_gc, interval: interval)
+    fg = Singed::Flamegraph::Stackprof.new(label: label, ignore_gc: ignore_gc, interval: interval)
     result = fg.record(&)
     fg.save
